align data

+++ include/Operations/Operation_ReluctorGM24x.h
@@ -1,7 +1,6 @@
 #include "Operations/IOperation.h"
 #include "Operations/Operation_DigitalPinRecord.h"
 #include "ReluctorResult.h"
-#include "Packed.h"
 #include "Interpolation.h"
 
 #ifndef OPERATION_RELUCTORGM24X_H
@@ -14,7 +13,7 @@ namespace OperationArchitecture
 		ReluctorResult Execute(Record *, uint32_t) override;
 		bool IsLongPulse(Record *, uint16_t frame);
 
-		static IOperationBase *Create(const void *config, unsigned int &sizeOut);
+		static IOperationBase *Create(const void *config, size_t &sizeOut);
 	};
 }
 #endif
