+++ src/Operations/Operation_ReluctorGM24x.cpp
@@ -455,7 +455,7 @@ namespace OperationArchitecture
 		return record->Frames[frame].Tick - record->Frames[frameMinus1].Tick > ticksPer7P5Degrees;
 	}
 
-	IOperationBase *Operation_ReluctorGM24x::Create(const void *config, unsigned int &sizeOut)
+	IOperationBase *Operation_ReluctorGM24x::Create(const void *config, size_t &sizeOut)
 	{
 		return new Operation_ReluctorGM24x();
 	}
